Store search metadata to database
I'm having second thoughts about the way search metadata is stored. I
made the json response to wait for the db to store the metadata first,
which means another thing to wait for (and make searches slower).

On the other hand, I could send the json response and store the metadata
in parallel. I'm not sure if it'll work, and if it did, one little
consequence would be a possible delay for the metadata to be stored
after the json response is sent. Probably not a big deal though.