Do not format the moment object
Formatting the moment object is awkward, since the model's timestamp
field is expecting a Date object, so what happens is, a moment object is
created, then formatted to a string, then coerced back to a Date, then
stored to the database. It results in the timestamp having '000' in its
milliseconds.